Accelerated tranpose using blocks

+++ src/common.h
@@ -88,13 +88,13 @@ return FAILURE; \
 
 #define FREADTEST(x, size, count, stream) \
 if(fread(x, size, count, stream) < count) { \
-fprintf(stderr, "read fewer items than expected (%d)\n", count); \
+fprintf(stderr, "read fewer items than expected (%lu)\n", (unsigned long)count); \
 return FAILURE; \
 }
 
 #define FWRITETEST(x, size, count, stream) \
 if(fwrite(x, size, count, stream) < count) { \
-fprintf(stderr, "wrote fewer items than expected (%d)\n", count); \
+fprintf(stderr, "wrote fewer items than expected (%lu)\n", (unsigned long)count); \
 return FAILURE; \
 }
